modem: move the MFSK waveform into modem/mfsk/
The DSP sat loose in modem/ alongside modem.c, framer.c and the backend
glue, which is the wrong shelf for it: modem/freedv/ already establishes that
a waveform's implementation lives in its own directory, and thirteen mfsk_*
files in the parent made it hard to see which sources are the modem layer and
which are one waveform's internals.

Pure renames -- git records all thirteen as R100, no content changed.  The
sources include each other by bare name and moved together, so only the
consumers needed touching: tests/Makefile and utils/Makefile gained the new
paths and -I../modem/mfsk, and utils/mfsk_puncture_sweep.c had the one
hardcoded "../modem/mfsk_ldpc.h" include.

Verified: full C build, unit suite 331 tests green with test_mfsk passing from
the new location, and mfsk_puncture_sweep links.  shortframe_sweep and
hail_suffix_sweep still fail to link (undefined `encode`, `run_ldpc_decoder`,
`ldpc_codes_list` -- their rules do not pull in libfreedvdata.a), but that is
pre-existing: checked out the branch before this change and reproduced the
identical failure there.  Worth fixing separately; this PR's description
claiming `make -C utils` clean is not accurate today.
