+++ utils/Makefile
@@ -1,7 +1,7 @@
 # Compiler and flags
 include ../config.mk
 
-CFLAGS = $(COMMON_CFLAGS) -Wextra -I../common -I../modem -I.. -I../modem/freedv
+CFLAGS = $(COMMON_CFLAGS) -Wextra -I../common -I../modem -I../modem/mfsk -I.. -I../modem/freedv
 LDFLAGS = -pthread
 
 ifeq ($(OS),Windows_NT)
@@ -65,18 +65,18 @@ rxcost_bench: rxcost_bench.c
 .PHONY: all clean run
 
 # MFSK waveform sweeps (pure DSP; no modem vtable, no ARQ)
-MFSK_DSP_SRC = ../modem/mfsk.c ../modem/mfsk_ofdm.c ../modem/mfsk_sync.c \
-               ../modem/mfsk_ldpc.c ../modem/mfsk_ldpc_1_16.c ../modem/mfsk_ldpc_2_16.c \
-               ../modem/mfsk_ldpc_3_16.c ../modem/mfsk_ldpc_5_16.c ../modem/mfsk_ldpc_8_16.c
+MFSK_DSP_SRC = ../modem/mfsk/mfsk.c ../modem/mfsk/mfsk_ofdm.c ../modem/mfsk/mfsk_sync.c \
+               ../modem/mfsk/mfsk_ldpc.c ../modem/mfsk/mfsk_ldpc_1_16.c ../modem/mfsk/mfsk_ldpc_2_16.c \
+               ../modem/mfsk/mfsk_ldpc_3_16.c ../modem/mfsk/mfsk_ldpc_5_16.c ../modem/mfsk/mfsk_ldpc_8_16.c
 
-shortframe_sweep: shortframe_sweep.c ../modem/mfsk.c
-	$(CC) $(CFLAGS) -I../modem -o $@ $^ -lm
+shortframe_sweep: shortframe_sweep.c ../modem/mfsk/mfsk.c
+	$(CC) $(CFLAGS) -I../modem -I../modem/mfsk -o $@ $^ -lm
 
-hail_suffix_sweep: hail_suffix_sweep.c ../modem/mfsk.c ../modem/mfsk_ofdm.c ../modem/mfsk_sync.c
-	$(CC) $(CFLAGS) -I../modem -o $@ $^ -lm
+hail_suffix_sweep: hail_suffix_sweep.c ../modem/mfsk/mfsk.c ../modem/mfsk/mfsk_ofdm.c ../modem/mfsk/mfsk_sync.c
+	$(CC) $(CFLAGS) -I../modem -I../modem/mfsk -o $@ $^ -lm
 
-mfsk_puncture_sweep: mfsk_puncture_sweep.c ../modem/mfsk_ldpc.c ../modem/mfsk_ldpc_8_16.c
-	$(CC) $(CFLAGS) -I../modem -o $@ $^ -lm
+mfsk_puncture_sweep: mfsk_puncture_sweep.c ../modem/mfsk/mfsk_ldpc.c ../modem/mfsk/mfsk_ldpc_8_16.c
+	$(CC) $(CFLAGS) -I../modem -I../modem/mfsk -o $@ $^ -lm
 # Broadcast a file to Mercury, or receive one, over the broadcast TCP/KISS port.
 BCAST_RQ_SRC = $(wildcard ../datalink_broadcast/raptorq/lib/*.c) \
                $(wildcard ../datalink_broadcast/raptorq/deps/obl/*.c)
